project/requirements.txt
openai 
requests 
pillow